refactor : move and delete page
