--- frontend/src/pages/RecommendationPage/RecommendationPage.module.css
@@ -1,30 +1,0 @@
-.recommendationPage {
-  padding: 20px;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.section {
-  margin-bottom: 40px;
-}
-
-.section h2 {
-  font-size: 24px;
-  margin-bottom: 10px;
-  color: #333;
-}
-
-.section p {
-  color: #666;
-  margin-bottom: 20px;
-}
-
-@media (max-width: 768px) {
-  .recommendationPage {
-    padding: 10px;
-  }
-
-  .section h2 {
-    font-size: 20px;
-  }
-}
