+++ frontend/src/pages/auth/SocialLoginCallbackPage/SocialLoginCallbackPage.tsx
